Added dataset validation messages
