+++ src/MNCD.Services/Impl/NetworkDataSetService.cs
@@ -24,7 +24,7 @@ namespace MNCD.Services.Impl
             _readerService = readerService;
         }
 
-        public async Task AddDataSet(string name, string content, FileType fileType)
+        public async Task<NetworkDataSet> AddDataSet(string name, string content, FileType fileType)
         {
             var hash = _hashService.GetHashFor(content);
             var existing = await GetNetworkDataSetByHash(hash);
@@ -35,7 +35,7 @@ namespace MNCD.Services.Impl
                 {
                     existing.Deleted = false;
                     await _ctx.SaveChangesAsync();
-                    return;
+                    return existing;
                 }
 
                 throw new ArgumentException($"DataSet already exists in the system with name {existing.Name}.");
@@ -56,6 +56,8 @@ namespace MNCD.Services.Impl
 
             await _ctx.DataSets.AddAsync(dataSet);
             await _ctx.SaveChangesAsync();
+
+            return dataSet;
         }
 
         public async Task DeleteDataSet(int id)
@@ -78,7 +80,7 @@ namespace MNCD.Services.Impl
                 .FirstOrDefaultAsync(d => d.Id == id);
         }
 
-        public async Task UpdateDataSet(int id, string name)
+        public async Task<NetworkDataSet> UpdateDataSet(int id, string name)
         {
             var dataSet = await _ctx
                 .DataSets
@@ -94,6 +96,8 @@ namespace MNCD.Services.Impl
             dataSet.Name = name;
 
             await _ctx.SaveChangesAsync();
+
+            return dataSet;
         }
 
         public async Task<List<NetworkDataSet>> GetDataSets()
@@ -109,32 +113,18 @@ namespace MNCD.Services.Impl
             return await _ctx.DataSets.FirstOrDefaultAsync(d => d.Hash == hash);
         }
 
-        private NetworkInfo GetNetworkInfo(string content, FileType fileType)
+        private NetworkInfo GetNetworkInfo(string content, FileType fileType) => fileType switch
         {
-            switch (fileType)
-            {
-                case FileType.MPX:
-                    return _readerService.ReadMPX(content);
-                case FileType.EdgeList:
-                    return _readerService.ReadEdgeList(content);
-                default:
-                    // TODO: custom exception
-                    throw new ArgumentException("File type is not supported.");
-            }
-        }
+            FileType.MPX => _readerService.ReadMPX(content),
+            FileType.EdgeList => _readerService.ReadEdgeList(content),
+            _ => throw new ArgumentException("File type is not supported."),// TODO: custom exception
+        };
 
-        private string GetEdgeList(string content, FileType fileType)
+        private string GetEdgeList(string content, FileType fileType) => fileType switch
         {
-            switch (fileType)
-            {
-                case FileType.MPX:
-                    return _readerService.ReadMPXToEdgeList(content);
-                case FileType.EdgeList:
-                    return _readerService.ReadEdgeListToString(content); ;
-                default:
-                    // TODO: custom exception
-                    throw new ArgumentException("File type is not supported.");
-            }
-        }
+            FileType.MPX => _readerService.ReadMPXToEdgeList(content),
+            FileType.EdgeList => _readerService.ReadEdgeListToString(content),
+            _ => throw new ArgumentException("File type is not supported."),// TODO: custom exception
+        };
     }
 }
