+++ src/MNCD.Web/ClientApp/src/components/analysis/flattening/basic-flattening.tsx
@@ -2,7 +2,7 @@ import React from "react";
 import { connect, ConnectedProps } from "react-redux";
 import { RootState } from "../../../store";
 import { Toggle, Stack } from "office-ui-fabric-react";
-import { updateFlatteningParameters } from "../../../slices/AnalysisSlice";
+import { updateFlatteningParameters } from "../../../slices/analysis-slice";
 
 class BasicFlattening extends React.Component<ReduxProps> {
   handleWeightEdgesChange = (_: any, checked?: boolean) => {
