+++ src/MNCD.Web/ClientApp/src/slices/index.ts
@@ -1,6 +1,6 @@
 import ThemeReducer from "./theme-slice";
-import AnalysisReducer from "./AnalysisSlice";
-import DataSetReducer from "./DataSetSlice";
+import AnalysisReducer from "./analysis-slice";
+import DataSetReducer from "./dataset-slice";
 import SessionReducer from "./session-slice";
 
 export { ThemeReducer, AnalysisReducer, DataSetReducer, SessionReducer };
